+++ main.go
@@ -32,9 +32,9 @@ func main() {
 		AtlantisVersion: atlantisVersion,
 	}
 	version := &cmd.VersionCmd{AtlantisVersion: atlantisVersion}
-	bootstrap := &cmd.BootstrapCmd{}
+	testdrive := &cmd.TestdriveCmd{}
 	cmd.RootCmd.AddCommand(server.Init())
 	cmd.RootCmd.AddCommand(version.Init())
-	cmd.RootCmd.AddCommand(bootstrap.Init())
+	cmd.RootCmd.AddCommand(testdrive.Init())
 	cmd.Execute()
 }
