+++ scripts/coverage.sh
@@ -6,7 +6,7 @@
 # with multiple packages, see https://code.google.com/p/go/issues/detail?id=6909
 #
 # Usage: coverage.sh packages...
-# Example: coverage.sh github.com/runatlantis/atlantis github.com/runatlantis/atlantis/bootstrap
+# Example: coverage.sh github.com/runatlantis/atlantis github.com/runatlantis/atlantis/testdrive
 #
 
 set -e
